Refactore & componentization
